User model - update

+++ AcademicHub/src/main/java/com/academichub/AcademicHub/interfaces/Features.java
@@ -7,17 +7,17 @@ public interface Features {
     Boolean register(String name, String lastName, String email, String password, Enum type);
     Boolean emailVerify(String email);
     Boolean insert(String name, String lastName, String email, String password, Enum type);
-    User getUser(Integer id);
-    Post getPost(Integer id);
+    User getUser(Long id);
+    Post getPost(Long id);
     Boolean updatePost(Post post,User user);
     Boolean updateUser(User user);
     Boolean deletePost(Post post,User user);
     Boolean deleteUser(User user);
     Boolean newPost(String description, T content);
-    Integer getLikes(Integer idPost);
-    void addLike(Integer idPost, Integer idUser);
-    List<Comment> getComments(Integer idPost);
-    void addComment(Integer idPost, Integer idUser, Comment comment);
+    Integer getLikes(Long idPost);
+    void addLike(Long idPost, Long idUser);
+    List<Comment> getComments(Long idPost);
+    void addComment(Long idPost, Long idUser, Comment comment);
     Boolean like();
     Boolean comment(Comment comment);
